amount: Move serde code to submodule
No changes other than moving the module code.
